4.21 token verification and user validation to blog deletion

+++ part4/blog-list/controllers/blogs.js
@@ -34,7 +34,22 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({error: 'token invalid'})
+  }
+
+  const user = await User.findById(decodedToken.id)
+  if (!user) {
+    return response.status(400).json({error: 'userId missing or not valid'})
+  }
   const id = request.params.id
+  const blog = await Blog.findById(id)
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({error: 'only the creator can delete this blog'})
+  }
+
   await Blog.findByIdAndDelete(id)
   response.status(204).end()
 })
